requirements.txt
numpy
scipy
Pillow
cython
matplotlib
scikit-image
tensorflow==1.4.0
keras>=2.0.8
opencv-python
h5py
imgaug
IPython[all]
bleach==1.5.0
